taskmate-api/mcp/requirements.txt
google-generativeai
fastapi
uvicorn
python-dotenv
websockets
